refactor: extract hotkey handler to src/content/hotkeyHandler.ts
- Move addHotkeyListener, hotkeyListener, seekFrameByKeyPressListener to new module
- Use dependency injection (initHotkeyHandler) for content.ts callbacks
- Create local Utils instance for biliBiliPlayerDanmakuInputBlur
- Re-export seekFrameByKeyPressListener from content.ts for backward compat
- Remove unused Keybind, keybindEquals, getFrameRate imports from content.ts
